backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
langchain==0.3.7
langchain-community==0.3.5
langchain-groq==0.2.1
langchain-huggingface==0.1.2
chromadb==0.5.23
sentence-transformers==3.3.1
pypdf==5.1.0
docx2txt==0.8
python-multipart==0.0.12
python-dotenv==1.0.1
pydantic==2.10.3
cryptography<42